# docs/09-参考/历史归档/工作模式云端接口-20260912.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Amie 工作模式接口联调契约",
    "version": "2026-09-12",
    "description": "当前实现的新增/改变接口子集，固定mock阶段；不是整站API清单或已接通云服务声明。个人记录仍由现有API持久化，模拟生成结果不入库。源码优先，详见同名Markdown。"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/reminders/scheduled/{id}": {
      "put": {
        "operationId": "updateScheduledReminder",
        "summary": "编辑本人提醒或将status改为paused/active",
        "x-source": "apps/api/src/routes/reminders.js",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "reminder"
                  ],
                  "properties": {
                    "reminder": {
                      "type": "object",
                      "description": "持久化后的ScheduledReminder记录"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "服务错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderUpdate"
              }
            }
          }
        },
        "description": "时间字段任一变化会重新计算nextFireAt；once需要date/time、weekly需要weekdays、monthly需要monthDay。含instruction的计划可保存，但mock阶段不会自动执行。"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
      "ReminderUpdate": {
        "type": "object",
        "required": [],
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "instruction": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "paused"
            ]
          },
          "freq": {
            "type": "string",
            "enum": [
              "once",
              "daily",
              "weekly",
              "monthly"
            ]
          },
          "time": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "weekdays": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "integer",
              "minimum": 0,
              "maximum": 6
            }
          },
          "monthDay": {
            "type": "integer",
            "minimum": 1,
            "maximum": 31
          }
        }
      },
...
    }
  }
}